BugFix for Lepton Isolation
It was found that the lepton isolation are changed to >1 instead of
<0.4 etc in the samples. This is because the Isolationvar got overwritten in
the IsoTrackFilter module, which calculated the isolation use PV tracks only.

To fix this problem, I add a new variable TrackIsolationVar to the Candidate
class to store the track isolation in the IsoTrackFilter module. However, the
TreeWriter for IsoTrack will still store the TrackIsolationVar into
Isolationvar.